Initial commit: original PGSA at SBU Constitution
The original text of the *Constitution* of the Physics Graduate Student
Association (PGSA) at Stony Brook University (SBU), last amended April
5, 2017 and last ratified "XX/XX/2017".

**Editors notes:**

At the time of this commit, the link to the "latest edition" of the
"Student Club and Organization Handbook" in Article VI, Section 1 is a
dead link.

Article VI, Section 4, Subsection C contains a formula that has been
altered for representation with Markdown.